pyproject.toml
[build-system]
requires = ["maturin>=1.5,<2.0"]
build-backend = "maturin"

[tool.poetry]
name = "gladoid-bevy"
version = "0.1.0"
description = ""
authors = ["Sergen Pekşen <peksensergen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
maturin = "^1.5.0"
dynaconf = "^3.2.4"

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"

[tool.maturin]
features = ["pyo3/extension-module"]
